mqtt_project_docker/gateway/requirements.txt
fastapi
uvicorn
websockets
grpcio
grpcio-tools
paho-mqtt
python-dateutil

